Move AIChatBot and Tab components to all-components
