+++ client/src/components/all-components/Tab/Tabs.js
