+++ client/src/components/exercise/lab13/pages/AIPanel.js
@@ -1,7 +1,7 @@
 import { React, useContext, useMemo, useState, useEffect, useRef } from "react";
 import { startExercise } from "src/reducers/lab2/actions";
 import { navigate } from "@reach/router";
-import { Tabs } from "../components/Tab/Tabs";
+import { Tabs } from "../../../all-components/Tab/Tabs";
 import ExerciseStateContext from "../Lab13Context";
 import { getTopicById } from "src/constants/lab13/BiasQuestionsConfig";
 import { content } from "src/constants/lab13/WikipediaContent";
